refactor: move ValidateHierarchicalRules, Kuadrant policy and hostnames to library
