+++ controllers/authpolicy_controller.go
@@ -16,7 +16,7 @@ import (
 	gatewayapiv1beta1 "sigs.k8s.io/gateway-api/apis/v1beta1"
 
 	api "github.com/kuadrant/kuadrant-operator/api/v1beta1"
-	"github.com/kuadrant/kuadrant-operator/pkg/common"
+	"github.com/kuadrant/kuadrant-operator/pkg/library/common"
 	"github.com/kuadrant/kuadrant-operator/pkg/reconcilers"
 )
