+++ controllers/ratelimitpolicy_controller.go
@@ -32,7 +32,7 @@ import (
 	gatewayapiv1beta1 "sigs.k8s.io/gateway-api/apis/v1beta1"
 
 	kuadrantv1beta2 "github.com/kuadrant/kuadrant-operator/api/v1beta2"
-	"github.com/kuadrant/kuadrant-operator/pkg/common"
+	"github.com/kuadrant/kuadrant-operator/pkg/library/common"
 	"github.com/kuadrant/kuadrant-operator/pkg/reconcilers"
 )
