+++ pkg/library/common/domains_test.go
@@ -1,0 +1,35 @@
+package common
+
+import "testing"
+
+func TestValidSubdomains(t *testing.T) {
+	testCases := []struct {
+		name             string
+		domains          []string
+		subdomains       []string
+		expected         bool
+		expectedHostname string
+	}{
+		{"nil", nil, nil, true, ""},
+		{"nil subdomains", []string{"*.example.com"}, nil, true, ""},
+		{"nil domains", nil, []string{"*.example.com"}, false, "*.example.com"},
+		{"dot matters", []string{"*.example.com"}, []string{"example.com"}, false, "example.com"},
+		{"dot matters2", []string{"example.com"}, []string{"*.example.com"}, false, "*.example.com"},
+		{"happy path", []string{"*.example.com", "*.net"}, []string{"*.other.net", "test.example.com"}, true, ""},
+		{"not all match", []string{"*.example.com", "*.net"}, []string{"*.other.com", "*.example.com"}, false, "*.other.com"},
+		{"wildcard in subdomains does not match", []string{"*.example.com", "*.net"}, []string{"*", "*.example.com"}, false, "*"},
+		{"wildcard in domains matches all", []string{"*", "*.net"}, []string{"*.net", "*.example.com"}, true, ""},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(subT *testing.T) {
+			valid, hostname := ValidSubdomains(tc.domains, tc.subdomains)
+			if valid != tc.expected {
+				subT.Errorf("expected (%t), got (%t)", tc.expected, valid)
+			}
+			if hostname != tc.expectedHostname {
+				subT.Errorf("expected hostname (%s), got (%s)", tc.expectedHostname, hostname)
+			}
+		})
+	}
+}
